+++ src/NeoComposeUnity/Tests/NeoActionEvaluatorTests.cs
@@ -556,9 +556,9 @@ namespace NeoCompose.Tests
         }
 
         [Test]
-        public void CurrentCompilerRevisionIsEleven()
+        public void CurrentCompilerRevisionIsTwelve()
         {
-            Assert.AreEqual(11, FunctionWithReturnType.CurrentCompilerRevision);
+            Assert.AreEqual(12, FunctionWithReturnType.CurrentCompilerRevision);
         }
 
         // -------------------------------------------------------------